pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gemini-ocr"
version = "0.1.0"
authors = [
    { name = "Tobias Sargeant", email = "tobias.sargeant@gmail.com" },
]
description = "Hybrid OCR with gemini and DocumentAI"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "google-cloud-storage",
    "google-genai",
    "google-cloud-documentai",
    "pydantic>=2.12.4",
    "pymupdf",
    "universal-pathlib>=0.3.5",
    "fsspec>=2025.10.0",
    "gcsfs>=2025.10.0",
    "seq-smith>=0.5.1",
    "flask>=3.1.2",
    "markdown>=3.10",
    "pygments>=2.17.2",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff>=0.14.6",
]

[project.urls]
"Homepage" = "https://github.com/folded/gemini-ocr"
"Bug Tracker" = "https://github.com/folded/gemini-ocr/issues"

[tool.ruff]
line-length = 120
target-version = "py311"
indent-width = 4

[tool.ruff.lint]
select = ["A", "B", "C",  "E", "F", "G", "I", "N", "Q", "S", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "ERA", "EXE", "ICN", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "UP", "YTT"]
ignore = [
    "ANN101",  # missing-type-self
    "PD011",   # pandas-use-of-dot-values (false positive)
]
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "ERA", "EXE", "FBT", "ICN", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "UP", "YTT"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "S101",     # asserts.
    "S102",     # exec().
    "PLR2004",  # magic value comparisons.
    "SLF001"    # private method access.
]
"docs/*" = [
    "A001"
]
"app.py" = [
    "ALL" # Ignore all checks for app.py as it contains embedded HTML/JS/CSS in a string.
]

[tool.mypy]
python_version = "3.11"

[tool.cibuildwheel]
build = "cp311-* cp312-* cp313-*"
